Add files via upload
Reduce the long duration of waiting in queues in shopping malls at the billing counter and provide a hassle-free, fast, efficient, and convenient solution.
Help visually impaired people at shopping malls to independently look out for the required products.

In product sensing and billing, once the product is sensed, it checks for availability of the product in stock. Then if available, it adds the product to bill and also displays the product details (Product name, MRP, expiry date)on the screen. If the product is sensed second time, it gets removed from the bill.
In voice assistance module, the details of the sensed product is read out and indication is given if the product is expired.

While during shopping, to prevent visually impaired people from bumping onto anything in the store, object detection technology is used. Ultrasonic sensors are fixed on front four corners of the trolley, so once an object is detected, it is indicated using a buzzer.